Indirection separating generators from its functions
